Create a marketplace type platform homepage. Scrape olx.ro for categories. We'll have automotive, etc.
Once the category is chosen, present user with filters (similar to OLX's).
Once filters are selected, instead of traditional ads, the user is dropped in a video timeline similar to TikTok's, where we'll present videos of the products they're after.

Requirements: 
- category homepage
- filters
- video timeline

We'll also need a "new ad" button that allows the user to upload pictures.
The pictures will be turned into a slideshow with random modern transitions using tools like ffmpeg and have on screen subtitles and voiceover if possible.
The voiceover should be AI generated based on the details the seller inputs, plus general information about the item. That can be added later